Abstracted out Signup logic inside of Header into its own component called <SignupEntryPt/>
